+++ CHANGES.rst
@@ -16,6 +16,9 @@ Changelog
 - Add @@export-incarichi view that allows to download a csv file with all Persona and their roles.
   [cekk]
 
+- Add tipologia_bando to summary serializer.
+  [cekk]
+
 6.0.21 (2023-10-30)
 -------------------